% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BC_dat}
\alias{BC_dat}
\title{Barcode distribution of an example experiment.}
\format{A S4 data object with the following slots:
\describe{
  \item{class}{sequence overview}
  \item{barcode read counts}{a data frame consisting of read counts and barcode sequences}
  \item{results dir}{path to a directory for any kind of results}
  \item{barcode backbone}{a string clarifying the barcode backbone structure}
  \item{label}{character string, used as label for file names etc.}
}}
\usage{
BC_dat
}
\description{
A dataset containing an example BCdat object which consists of 98 barcode sequences and with no error correction yet.
}
\details{
BC_dat:
}
\keyword{datasets}
