% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BCdata-class.R
\docType{class}
\name{BCdat-class}
\alias{BCdat-class}
\alias{BCdat}
\title{BCdat class.}
\description{
BCdat class.
}
\section{Slots}{

\describe{
\item{\code{reads}}{data.frame containing barcode sequences and their corresponding read counts.}

\item{\code{results_dir}}{character string of the working directory path.}

\item{\code{label}}{character string identifying the particular experiment (will be part of the names of any file created).}

\item{\code{BC_backbone}}{character string of the used barcode design (also called barcode backbone).}
}}

