% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BCdata-class.R
\name{checkBarcodeData}
\alias{checkBarcodeData}
\title{checkBarcodeData}
\usage{
checkBarcodeData(object)
}
\arguments{
\item{object}{a BCdat object.}
}
\description{
Checks data slots of BCdat object for correctness.
}
