% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{com_pair}
\alias{com_pair}
\title{Compairing two BCdat Objects}
\usage{
com_pair(BC_dat1 = NULL, BC_dat2 = NULL)
}
\arguments{
\item{BC_dat1}{the first BCdat object.}

\item{BC_dat2}{the second BCdat object.}
}
\value{
a list containing the shared and the unqiue barcodes.
}
\description{
Compairing two BCdat Objects
}
