% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{.createPatternFile}
\alias{.createPatternFile}
\title{Internal function}
\usage{
.createPatternFile(bc_backbone, patterns_file)
}
\arguments{
\item{bc_backbone}{a character string (barcode pattern).}

\item{patterns_file}{a character string (file name)}
}
\description{
Creates a search file for a command line grep search.
}
