% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{.generateColors}
\alias{.generateColors}
\title{Color list generation}
\usage{
.generateColors(minHD, type = "rainbow", alpha = 1)
}
\arguments{
\item{minHD}{a numeric vector of all the minimum hamming distances.}

\item{type}{a character string. Possible Values are "rainbow", "heat.colors", "topo.colors", "greens", "wild".}

\item{alpha}{a numeric value between 0 and 1, modifies colour transparency.}
}
\description{
Generates a collection of colors for a list of barcodes based on their identified minimum hamming distances.
}
