% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{.getBarcodeFilter}
\alias{.getBarcodeFilter}
\title{Internal function}
\usage{
.getBarcodeFilter(wobble_pos)
}
\arguments{
\item{wobble_pos}{a character string.}
}
\description{
Identifies the barcode positions within the barcode backbone and generates a awk command.
}
