% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{.getDiagonalIndex}
\alias{.getDiagonalIndex}
\title{Index Generation}
\usage{
.getDiagonalIndex(n)
}
\arguments{
\item{n}{an integer indicating the size of the resulting index matrix.}
}
\value{
a locigal matrix of size \code{n} x \code{n}
}
\description{
Generates a matrix index to create a square triangular matrix.
}
