% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{.getMinDist}
\alias{.getMinDist}
\title{Distance calculation}
\usage{
.getMinDist(BC_dat, ori_BCs, m = "hamming")
}
\arguments{
\item{BC_dat}{a BCdat object}

\item{ori_BCs}{a character vector containing barcodes to which the minimal hamming distance will be calculated.}

\item{m}{a character string, Method for distance calculation, default value is Hamming distance. Possible values
are "osa", "lv", "dl", "hamming", "lcs", "qgram", "cosine", "jaccard", "jw", "soundex" (see stringdist function
of the stringdist-package for more information).}
}
\description{
Calculates the minimum distance to a set of predefined barcodes for a given list of barcode.
}
