% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{.getWobblePos}
\alias{.getWobblePos}
\title{getWobblePos}
\usage{
.getWobblePos(bc_backbone = "")
}
\arguments{
\item{bc_backbone}{a character vector.}
}
\description{
Extracts barcode positions.
}
