% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{.hex2rgbColor}
\alias{.hex2rgbColor}
\title{Converts hex colors into gephi usable rgb colors}
\usage{
.hex2rgbColor(colrs)
}
\arguments{
\item{colrs}{a character vector containing a list of hex colors}
}
\value{
a color vector.
}
\description{
Converts hex colors into gephi usable rgb colors
}
