% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{.revComp}
\alias{.revComp}
\title{DNA string manipulation}
\usage{
.revComp(seq_dat)
}
\arguments{
\item{seq_dat}{a character vector containing DNA sequences}
}
\description{
Converts a vector of character strings (DNA sequences) into its reverse complement.
}
