% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{.revComp_EqLength}
\alias{.revComp_EqLength}
\title{DNA string manipulation for equal string sizes}
\usage{
.revComp_EqLength(seq_dat, word_length)
}
\arguments{
\item{seq_dat}{a character vector.}

\item{word_length}{an integer giving the word length.}
}
\description{
Converts a vector of equally long character strings into its reverse complement.
}
