% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{.revComp_UneqLength}
\alias{.revComp_UneqLength}
\title{DNA string manipulation for unequal string sizes}
\usage{
.revComp_UneqLength(seq_dat)
}
\arguments{
\item{seq_dat}{A character vector.}
}
\description{
Converts a vector of unequally long character strings into the reverse complement.
}
