% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{.testDirIdentifier}
\alias{.testDirIdentifier}
\title{Internal function}
\usage{
.testDirIdentifier(s)
}
\arguments{
\item{s}{a character string.}
}
\description{
Checks directory paths for correctness and if nessesary corrects them.
}
