% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{error_correction_clustered_HDs}
\alias{error_correction_clustered_HDs}
\title{Clustered HD Plot}
\usage{
error_correction_clustered_HDs(datEC, size = 0.75)
}
\arguments{
\item{datEC}{a BC_dat object, returned by the error_correction function with the EC_analysis parameter set to TRUE.}

\item{size}{a numeric value, specifying the dot size.}
}
\value{
a ggplot2 object.
}
\description{
This function will create a jitter plot displaying the maximal distances within each of the barcode sequence clusters.
}
