% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{genBaRcode_app}
\alias{genBaRcode_app}
\title{Shiny App}
\usage{
genBaRcode_app(dat_dir = system.file("extdata", package = "genBaRcode"))
}
\arguments{
\item{dat_dir}{a character string, identifying the path to one or more fast(q) files which shall be analysed, default is the
path to the package inherent example fastq file}
}
\description{
Launches the corresponding shiny app.
}
