% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{generateKirchenplotRCdis}
\alias{generateKirchenplotRCdis}
\title{Plotting a Kirchenplot}
\usage{
generateKirchenplotRCdis(BC_dat, b = 30, show_it = FALSE)
}
\arguments{
\item{BC_dat}{a BCdat object.}

\item{b}{an integer value, defining the number of bins.}

\item{show_it}{a logical vaue. If TRUE, the respective values are printed on the console?}
}
\value{
ggplot2 object
}
\description{
Generates a barplot visualising the read count distribution.
}
\examples{
data(BC_dat)
generateKirchenplotRCdis <- function(BC_dat, b = 10, show_it = TRUE)
}
