% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{generateTimeSeriesData}
\alias{generateTimeSeriesData}
\title{Generating Time Series Data Object}
\usage{
generateTimeSeriesData(BC_dat_list)
}
\arguments{
\item{BC_dat_list}{a list of BCdat objects.}
}
\value{
a data.frame containing every identified barcode and its read count per time point/measurement.
}
\description{
Generates a matrix containing barcodes sequences as rows and consecutive measurements at columns. It serves as
the necessary data object for the plotting function `plotTimeSeries`.
}
