% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{getBackbone}
\alias{getBackbone}
\title{Predefined Barcode Backbone Sequences}
\usage{
getBackbone(id = NULL)
}
\arguments{
\item{id}{an integer or character value in order to choose a specific backbone.}
}
\value{
a character string.
}
\description{
allows one to choose between predefined backbone sequences. Excecution of the function without any parameter value will display all
available backbone sequences. The id parameter will accept the name of the backbone or the rownumber of the shown selection.
}
