% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BCdata-class-methods.R
\name{getLabel}
\alias{getLabel}
\title{Accessing the Label slot of a BCdat objects.}
\usage{
getLabel(object)
}
\arguments{
\item{object}{a BCdat object.}
}
\value{
A character string.
}
\description{
Accessing the Label slot of a BCdat objects.
}
\examples{
data(BC_dat)
getLabel(BC_dat)

}
