% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BCdata-class-methods.R
\name{getReads}
\alias{getReads}
\title{Accessing the Read-Count slot of a BCdat objects.}
\usage{
getReads(object)
}
\arguments{
\item{object}{a BCdat object.}
}
\value{
A data.frame containing the read count table of the object paramter.
}
\description{
Accessing the Read-Count slot of a BCdat objects.
}
\examples{
data(BC_dat)
getReads(BC_dat)

}
