% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BCdata-class-methods.R
\name{getResultsDir}
\alias{getResultsDir}
\title{Accessing the Results Directory slot of a BCdat objects.}
\usage{
getResultsDir(object)
}
\arguments{
\item{object}{a BCdat object.}
}
\value{
A character string.
}
\description{
Accessing the Results Directory slot of a BCdat objects.
}
\examples{
data(BC_dat)
getResultsDir(BC_dat)

}
