% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_data_processing.R
\name{hybridsIdentification}
\alias{hybridsIdentification}
\title{Identifies hybrid barcodes}
\usage{
hybridsIdentification(dat, min_seq_length = 2)
}
\arguments{
\item{dat}{a character vector containing barcode sequences or a BCdat object.}

\item{min_seq_length}{a positive integer value indicating the minimal length of the two barcodes which give rise to a hybrid barcode.}
}
\value{
a hybrid-free frequency table of barcode sequences
}
\description{
Experimental function to identify hybrid barcodes which can occure due to unfinished synthesis of a template
in-between PCR cycles.
}
\examples{
data(BC_dat)
hybridsIdentification(BC_dat, min_seq_length = 2)
}
