% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data2.R
\docType{data}
\name{ori_BCs}
\alias{ori_BCs}
\title{List of Barcodes.}
\format{a character vector:}
\usage{
ori_BCs
}
\description{
An object called ori_BCs containing specific predefined barcode sequences.
}
\details{
ori_BCs:
}
\keyword{datasets}
