% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plotHammDistIgraph}
\alias{plotHammDistIgraph}
\title{Plotting a Hamming Distance Network}
\usage{
plotHammDistIgraph(BC_dat, minHD = 1, loga = TRUE, ori_BCs,
  threeD = FALSE, complete = TRUE, col_type = "rainbow")
}
\arguments{
\item{BC_dat}{a BCdat object.}

\item{minHD}{an integer value representing the maximal hamming distance value for which the graph will
contain edges.}

\item{loga}{a logical value, indicating the use or non-use of logarithmic read count values.}

\item{ori_BCs}{a vector of character strings containing the initial barcode sequences.}

\item{threeD}{a logical value to chose between 2D and 3D visualisation.}

\item{complete}{a logical value. If TRUE, every node will have at least one edge.}

\item{col_type}{a character sting, choosing one of the available color palettes.}
}
\value{
an igraph object.
}
\description{
plotHammDistIgraph will create a graph-like visualisation (ripple plot) of the corresponding barcode sequences
and their similarity based on the igraph package. The nodes represent the barcode sequences and their
respective size reflects the corresponding read counts. Edges between nodes indicate a hamming distance between two barcodes
of maximal \code{minHD}.
If \code{ori_BCs} is provided the node color also refelects the hamming distance of a particular barcode to one of the initial
barcodes.
}
\examples{
data(BC_dat)
plotHammDistIgraph(BC_dat, minHD = 1, loga = TRUE, ori_BCs, threeD = FALSE,
complete = TRUE, col_type = "rainbow")
}
