% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plotNucFrequency}
\alias{plotNucFrequency}
\title{Plotting Nucleotide Frequency}
\usage{
plotNucFrequency(source_dir, file_name)
}
\arguments{
\item{source_dir}{a character string containing the path to the sequencing file.}

\item{file_name}{a character string containng the name of the sequencing file.}
}
\value{
a ggplot2 object.
}
\description{
Creates a plot visualising the nucleotide frequency within the entire fastq file.
}
