% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plotQualityScoreDis}
\alias{plotQualityScoreDis}
\title{Plotting Quality Score Distribution}
\usage{
plotQualityScoreDis(source_dir, file_name, type)
}
\arguments{
\item{source_dir}{a character string of the path to the source directory.}

\item{file_name}{a character string of the file name.}

\item{type}{a character string, possible values are "mean" and "median".}
}
\value{
a ggplot2 object.
}
\description{
Creates a plot of the quality values accommodated by the fastq file.
}
\examples{

\dontrun{

source_dir <- system.file("extdata", package = "genBaRcode")

plotQualityScoreDis(source_dir, file_name = "test_data.fastq", type = "mean")

}

}
