% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plotQualityScorePerCycle}
\alias{plotQualityScorePerCycle}
\title{Plotting Quality Score per Cycle}
\usage{
plotQualityScorePerCycle(source_dir, file_name)
}
\arguments{
\item{source_dir}{a character string containing the path to the sequencing file.}

\item{file_name}{a character string containng the name of the sequencing file.}
}
\value{
a ggplot2 object.
}
\description{
Visualises the mean, median, 25% and 75% quantile of the quality score per sequencing cycle.
}
