% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plotReadFrequencies}
\alias{plotReadFrequencies}
\title{Plotting a Barplot}
\usage{
plotReadFrequencies(BC_dat, b = 30, show_it = FALSE, log = FALSE)
}
\arguments{
\item{BC_dat}{a BCdat object.}

\item{b}{an integer value, defining the number of bins.}

\item{show_it}{a logical vaue. If TRUE, the respective values are printed on the console?}

\item{log}{a logical vaue. If TRUE, the y-axis will be on a log scale.}
}
\value{
ggplot2 object
}
\description{
Generates a barplot visualising the abundances of unique read count frequencies.
}
