% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plotSeqLogo}
\alias{plotSeqLogo}
\title{Plots a sequence logo}
\usage{
plotSeqLogo(BC_dat, colrs = NULL)
}
\arguments{
\item{BC_dat}{a chatacter vector or BCdat object containing the respective sequences}

\item{colrs}{a character vector containing the desired colors for the nucleotides A, T, C, G and N (in that order)}
}
\value{
a ggplot2 object
}
\description{
Plots a sequence logo
}
