% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_data_processing.R
\name{prepareDatObject}
\alias{prepareDatObject}
\title{Data Object Preparation}
\usage{
prepareDatObject(dat, results_dir, label, bc_backbone, min_reads, save_it)
}
\arguments{
\item{dat}{a tbl_df object (e.g. created by dplyr::count)}

\item{results_dir}{a character string which contains the path to the results directory.}

\item{label}{a character string which serves as a label for every kind of created output file.}

\item{bc_backbone}{a character string describing the barcode design, variable positions have to be marked with the letter 'N'.}

\item{min_reads}{positive integer value, all extracted barcode sequences with a read count smaller than min_reads will be excluded from the results}

\item{save_it}{a logical value. If TRUE, the raw data will be saved as a csv-file.}
}
\value{
a BCdat object.
}
\description{
generates BCdat object after barcode backbone identification.
}
