% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{readBCdat}
\alias{readBCdat}
\title{Data Input}
\usage{
readBCdat(path = "./", label = "", BC_backbone = "", file_name,
  s = ";")
}
\arguments{
\item{path}{a character string containing the path to a saved read count table (two columns containing read counts and barcode
sequences).}

\item{label}{a character string containing a label of the data set.}

\item{BC_backbone}{a character string containing the barcode structure information.}

\item{file_name}{a character string containing the name of the file to read in.}

\item{s}{a character value, identifying the column separating char.}
}
\value{
a BCdat object.
}
\description{
Reads a data table (csv-file) and returns a BCdat objects.
}
