% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BCdata-class-methods.R
\name{setBackbone}
\alias{setBackbone}
\title{Replacing the Barcode Backbone slot of a BCdat objects.}
\usage{
setBackbone(object, value)
}
\arguments{
\item{object}{a BCdat object.}

\item{value}{a character string consisting of exclusively IUPAC-nucleotide-code conform letters.}
}
\value{
a BCdat object.
}
\description{
Replacing the Barcode Backbone slot of a BCdat objects.
}
\examples{
data(BC_dat)
new_backbone <- getBackboneSelection("BC32-T-Sapphire")
BC_dat_alt <- setBackbone(BC_dat, new_backbone)

}
