% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BCdata-class-methods.R
\name{setLabel}
\alias{setLabel}
\title{Replacing the Label slot of a BCdat objects.}
\usage{
setLabel(object, value)
}
\arguments{
\item{object}{a BCdat object.}

\item{value}{a character string.}
}
\value{
a BCdat object.
}
\description{
Replacing the Label slot of a BCdat objects.
}
\examples{
data(BC_dat)
new_label <- "foo-bar"
BC_dat_alt <- setLabel(BC_dat, new_label)

}
