% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BCdata-class-methods.R
\name{setResultsDir}
\alias{setResultsDir}
\title{Replacing the Results Directory slot of a BCdat objects.}
\usage{
setResultsDir(object, value)
}
\arguments{
\item{object}{a BCdat object.}

\item{value}{a character string of an existing path.}
}
\value{
a BCdat object.
}
\description{
Replacing the Results Directory slot of a BCdat objects.
}
\examples{
data(BC_dat)
new_path <- getwd()
BC_dat_alt <- setResultsDir(BC_dat, new_path)

}
