% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BCdata-class-methods.R
\name{getBackbone}
\alias{getBackbone}
\title{Accessing the Barcode Backbone slot of a BCdat objects.}
\usage{
getBackbone(object)
}
\arguments{
\item{object}{a BCdat object.}
}
\value{
A character string.
}
\description{
Accessing the Barcode Backbone slot of a BCdat objects.
}
\examples{
data(BC_dat)
getBackbone(BC_dat)

}
